file reorganization
Reorganization of the package and file names according to PEP standards.
