+++ bal/Wallet.py
@@ -1,5 +1,5 @@
 from ecdsa import SigningKey, SECP256k1
-from Transaction import get_transaction_id, sign_tx_in, new_transaction, new_tx_in, new_tx_out, new_unspent_tx_out, find_unspent_tx_out
+from bal.transaction import get_transaction_id, sign_tx_in, new_transaction, new_tx_in, new_tx_out, new_unspent_tx_out, find_unspent_tx_out
 import numbers
 import os
 import json
