+++ bal/variant/base_blockchain.py
@@ -10,10 +10,10 @@ from functional import seq
 import numbers
 import threading
 
-from p2p import P2P
-from Transaction import new_coinbase_transaction, is_valid_address, process_transactions, new_transaction, COINBASE_AMOUNT
-from TransactionPool import TransactionPool
-from Wallet import create_transaction, find_unspent_tx_outs, get_balance, get_private_from_wallet, get_public_from_wallet
+from bal.p2p import P2P
+from bal.transaction import new_coinbase_transaction, is_valid_address, process_transactions, new_transaction, COINBASE_AMOUNT
+from bal.transaction_pool import TransactionPool
+from bal.wallet import create_transaction, find_unspent_tx_outs, get_balance, get_private_from_wallet, get_public_from_wallet
 
 DIFFICULTY_ADJUSTMENT_INTERVAL = 16 # block number
 BLOCK_GENERATION_INTERVAL = 2 # in seconds
