+++ bal/variant/pow_blockchain.py
@@ -1,4 +1,4 @@
-from BaseBlockchain import BaseBlockchain
+from bal.variant.base_blockchain import BaseBlockchain
 import numbers
 from time import time
 import hashlib
