+++ simulation/complex_networks/bcmn_real_world_simulation.py
@@ -13,7 +13,7 @@ from mininet.util import specialClass
 from mininet.link import TCLink
 from mininet.term import makeTerms
 from bal.bcnode import POWNode, POSNode
-from bal.BaseBlockchain import BLOCK_GENERATION_INTERVAL
+from bal.variant.base_blockchain import BLOCK_GENERATION_INTERVAL
 from collections import defaultdict
 import random
 import itertools
